Change parameter type of ValueStream.skip from int to long
More than 2G positions can legitimately be skipped at once

+++ presto-orc/src/main/java/com/facebook/presto/orc/stream/ByteArrayStream.java
@@ -59,7 +59,7 @@ public class ByteArrayStream
     }
 
     @Override
-    public void skip(int skipSize)
+    public void skip(long skipSize)
             throws IOException
     {
         skipFully(inputStream, skipSize);
